% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_starred_cran.R
\name{display_most_starred}
\alias{display_most_starred}
\title{Display the most starred R Github Repositories}
\usage{
display_most_starred(n = 10)
}
\arguments{
\item{n}{the number of most starred Github R repositories to fetch.
Defaults to 10.}
}
\value{
a character vector of the most starred R repositories
}
\description{
Display the most starred R Github Repositories
}
