% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_starred_cran.R
\name{display_starred}
\alias{display_starred}
\title{Display User's Github Starred Repositories}
\usage{
display_starred(github_user, n = 5, onlyR = FALSE)
}
\arguments{
\item{github_user}{the Github user name to look for}

\item{n}{the number of the last starred repositories. Defaults to 5 in which case
it will return the last 5 starred repositories. Note that if the 'onlyR' parameter
is set to TRUE, you might get a lower number of starred repos due to filtering R from
all the other languages.}

\item{onlyR}{Logical, whether to fetch only R repositories, Default to FALSE}
}
\value{
A character vector of starred Github repositories
}
\description{
Display User's Github Starred Repositories
}
