% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresh_start.R
\name{fresh_start}
\alias{fresh_start}
\title{Remove all the installed R packages}
\usage{
fresh_start()
}
\value{
called for the side effect of removing all installed packages
}
\description{
Remove all the installed R packages
}
\examples{

\dontrun{

# DANGER: THE FUNCTION REMOVES ALL THE PACKAGES

fresh_start()

}
}
