% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_starred_cran.R
\name{install_most_starred}
\alias{install_most_starred}
\title{Install the most starred CRAN packages}
\usage{
install_most_starred(n = 10)
}
\arguments{
\item{n}{the most starred starred CRAN packages. Defaults to 10.
in this case the function will look at the 10 most starred R repo and install them
if available on CRAN.}
}
\value{
called for the side effect of installing most starred CRAN packages
}
\description{
Install the most starred CRAN packages
}
