% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_starred_cran.R
\name{install_starred_github}
\alias{install_starred_github}
\title{Install Github Starred Packages from Github}
\usage{
install_starred_github(github_user, n = 5, upgrade = "never")
}
\arguments{
\item{github_user}{the Github user name to look for}

\item{n}{the last 'n' starred repositories. Defaults to 5, in which case it will look for
the last 5 starred repositories, filter the R repos and install them}

\item{upgrade}{whether to upgrade out of date packages. You can choose from
'always' or 'never'. Defaults to 'never'. For more info,
see <install_github()> from the 'remote' package.}
}
\value{
called for the side effect of installing the Github starred repositories
}
\description{
installs the Github starred repositories from Github and not from CRAN.
}
