% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_packages.R
\name{latest_packages}
\alias{latest_packages}
\title{Displaying the latest installed R packages}
\usage{
latest_packages(n)
}
\arguments{
\item{n}{the number of the last installed packages to display (n = 1) will return the last installed package}
}
\value{
a data frame
}
\description{
Displaying the latest installed R packages
}
\examples{
\dontrun{

# Displaying the last 10 installed packages
latest_packages(10)

}
}
