% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_last_package.R
\name{rm_last_package}
\alias{rm_last_package}
\title{Remove the latest installed R package}
\usage{
rm_last_package()
}
\value{
called for the side effect of removing the last installed package
}
\description{
Remove the latest installed R package
}
\examples{
\dontrun{

# Removing the last installed package
rm_last_package()

}


}
