% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_today_packages.R
\name{rm_today_packages}
\alias{rm_today_packages}
\title{Remove the packages installed in the current day}
\usage{
rm_today_packages(lib = .libPaths())
}
\arguments{
\item{lib}{a character vector giving the library directories. Defaults to the first element in .libPaths()}
}
\value{
called for the side effect of removing the today installed packages
}
\description{
Remove the packages installed in the current day
}
\examples{

\dontrun{

# Removing the packages installed today

rm_today_packages()
}
}
