% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_yesterday_packages.R
\name{rm_yesterday_packages}
\alias{rm_yesterday_packages}
\title{Remove the packages installed yesterday}
\usage{
rm_yesterday_packages()
}
\value{
called for the side effect of removing the yesterday installed packages
}
\description{
Remove the packages installed yesterday
}
\examples{

\dontrun{
# Removing the packages installed yesterday
rm_yesterday_packages()
}
}
