% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/since_packages.R
\name{since_packages}
\alias{since_packages}
\title{Displays installed packages according to a specific date}
\usage{
since_packages(date, position)
}
\arguments{
\item{date}{the date of interest in yyyy-mm-dd format}

\item{position}{takes three arguments "at", "before" or "after". "at" displays the packages installed at the chosen date, "before" before that date and "after" after that date)}
}
\value{
a character vector
}
\description{
Displays installed packages according to a specific date
}
\examples{
\dontrun{
# Displaying the packages installed today
since_packages(date = Sys.Date(), position = "at")
}
}
