% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today_packages.R
\name{today_packages}
\alias{today_packages}
\title{Displays the packages installed in the current day}
\usage{
today_packages()
}
\value{
a character vector
}
\description{
Displays the packages installed in the current day
}
\examples{
\dontrun{
# Displaying the packages installed today
today_packages()
}
}
