% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesterday_packages.R
\name{yesterday_packages}
\alias{yesterday_packages}
\title{Displays the packages installed yesterday}
\usage{
yesterday_packages()
}
\value{
a character vector
}
\description{
Displays the packages installed yesterday
}
\examples{
\dontrun{
# Displaying the packages installed yesterday
yesterday_packages()
}
}
