% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_latest_packages.R
\name{rm_latest_packages}
\alias{rm_latest_packages}
\title{Remove the n latest installed R packages}
\usage{
rm_latest_packages(n = 1, lib = .libPaths())
}
\arguments{
\item{n}{the last number of installed packages to remove. Default to 1 for the last installed package}

\item{lib}{a character vector giving the library directories. Defaults to the first element in .libPaths()}
}
\value{
called for the side effect of removing the n latest installed packages
}
\description{
Remove the n latest installed R packages
}
\examples{

\dontrun{

# Removing the last 10 installed packages

rm_latest_packages(n = 10)

}
}
