% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_latest_packages.R
\name{rm_latest_packages}
\alias{rm_latest_packages}
\title{Remove a specific number of the latest installed R package}
\usage{
rm_latest_packages(n)
}
\arguments{
\item{n}{the last number of installed packages to remove}
}
\value{
called for the side effect of removing the n latest installed packages
}
\description{
Remove a specific number of the latest installed R package
}
\examples{

\dontrun{

# Removing the last 10 installed packages

rm_latest_packages(n = 10)

}
}
